https://github.com/DASISH/qddt-client/issues/463

+++ src/main/java/no/nsd/qddt/domain/controlconstruct/ControlConstructServiceImpl.java
@@ -149,7 +149,7 @@ class ControlConstructServiceImpl implements ControlConstructService {
         instance.populateControlConstructInstructions();
 
         if(instance.isBasedOn()) {
-            Integer rev= auditService.findLastChange(instance.getId()).getRevisionNumber();
+            Long rev= auditService.findLastChange(instance.getId()).getRevisionNumber().longValue();
             instance.makeNewCopy(rev);
         } else if (instance.isNewCopy()) {
             instance.makeNewCopy(null);
