+++ src/main/java/no/nsd/qddt/domain/questionItem/QuestionItemServiceImpl.java
@@ -176,7 +176,7 @@ class QuestionItemServiceImpl implements QuestionItemService {
     private QuestionItem prePersistProcessing(QuestionItem instance){
 
         if(instance.isBasedOn()) {
-            Integer rev= auditService.findLastChange(instance.getId()).getRevisionNumber();
+            Long rev= auditService.findLastChange(instance.getId()).getRevisionNumber().longValue();
             instance.makeNewCopy(rev);
         } else if (instance.isNewCopy()){
             instance.makeNewCopy(null);
